ds-application/requirements.txt
pandas
requests
gunicorn
python-dotenv
python-dateutil
flask
folium